+++ rust/kcl-lib/src/lib.rs
@@ -93,11 +93,11 @@ pub use lsp::{
     kcl::{Backend as KclLspBackend, Server as KclLspServerSubCommand},
 };
 pub use modules::ModuleId;
-#[cfg(not(target_arch = "wasm32"))]
-pub use parsing::ast::recast_dir;
 pub use parsing::ast::{modify::modify_ast_for_sketch, types::FormatOptions};
 pub use settings::types::{project::ProjectConfiguration, Configuration, UnitLength};
 pub use source_range::SourceRange;
+#[cfg(not(target_arch = "wasm32"))]
+pub use unparser::recast_dir;
 
 // Rather than make executor public and make lots of it pub(crate), just re-export into a new module.
 // Ideally we wouldn't export these things at all, they should only be used for testing.
